Write notes for a Sudoku solver
I've heard that one of the problems with solving Sudoku on a computer is
that the brute-force algorithm is so fast that the overhead required to
use fancier algorithms slows things down. I have an idea of an algorithm
that should allow for more optimal brute-force tactics than array
scanning for conflicts, trying to fill every square that is not already
filled with an initial hint.